+++ docs/CHANGELOG.md
@@ -26,6 +26,12 @@ Check the php manual for help making your desired string. ([<https://www.php.net
   * podcast: get a podcast by id
   * podcast_episodes: get a list of podcast_episodes you can access
   * podcast_episode: get a podcast_episode by id
+  * podcast_create: create a podcast
+  * podcast_edit: edit an existing podcast
+  * podcast_delete: delete an existing podcast
+  * update_podcast: sync and download new episodes
+  * catalogs: get all the catalogs
+  * catalog: get a catalog by id
 * NEW db options
   * cron_cache: Speed up the interface by allowing background caching of data
   * show_skipped_times: Add "# skipped" to the ui. (disabled by default)